+++ openvkl/drivers/ispc/volume/vdb/VdbLeafAccessObserver.h
@@ -3,20 +3,25 @@
 
 #pragma once
 
-#include "../common/Observer.h"
+#include "../../common/Allocator.h"
+#include "../../observer/Observer.h"
+#include "../../observer/ObserverRegistry.h"
 #include "openvkl/ispc_cpp_interop.h"
 
 namespace openvkl {
   namespace ispc_driver {
 
+    template <int W>
+    class VdbSampler;
+    struct VdbGrid;
+
     /*
      * The leaf access observer simply wraps the buffer allocated by VdbVolume.
      */
-    struct VdbLeafAccessObserver : public Observer
+    template <int W>
+    struct VdbLeafAccessObserver : public Observer<W>
     {
-      VdbLeafAccessObserver(ManagedObject &target,
-                            size_t size,
-                            const vkl_uint32 *accessBuffer);
+      VdbLeafAccessObserver(VdbSampler<W> &target, const VdbGrid &grid);
 
       VdbLeafAccessObserver(VdbLeafAccessObserver &&) = delete;
       VdbLeafAccessObserver &operator=(VdbLeafAccessObserver &&) = delete;
@@ -31,9 +36,12 @@ namespace openvkl {
       size_t getNumElements() const override;
 
      private:
-      ManagedObject *target{nullptr};
+      ObserverRegistry<W> &getRegistry();
+
+     private:
+      Allocator allocator;
       size_t size{0};
-      const vkl_uint32 *accessBuffer{nullptr};
+      vkl_uint32 *accessBuffer{nullptr};
     };
 
   }  // namespace ispc_driver