+++ openvkl/drivers/ispc/volume/vdb/VdbSampler.cpp
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "VdbSampler.h"
+#include "VdbLeafAccessObserver.h"
 #include "VdbSampler_ispc.h"
 #include "VdbVolume.h"
 
@@ -9,10 +10,9 @@ namespace openvkl {
   namespace ispc_driver {
 
     template <int W>
-    VdbSampler<W>::VdbSampler(const VdbGrid *grid,
-                              const VdbSampleConfig &defaultConfig)
-        : grid(grid), config(defaultConfig)
+    VdbSampler<W>::VdbSampler(VdbVolume<W> &volume) : volume(&volume)
     {
+      config.leafAccessObservers = leafAccessObservers.getIE();
     }
 
     template <int W>
@@ -21,19 +21,21 @@ namespace openvkl {
     template <int W>
     void VdbSampler<W>::commit()
     {
-      config.filter =
-          (VKLFilter)this->template getParam<int>("filter", config.filter);
+      const VdbSampleConfig &globalConfig = volume->getSampleConfig();
+      config.filter = (VKLFilter)this->template getParam<int>(
+          "filter", globalConfig.filter);
 
       // Note: We fall back to the sampler object filter parameter if it is set.
       //       This enables users to specify *only* the field filter override.
-      //       This does mean that users must set the gradientFilter explicitly 
+      //       This does mean that users must set the gradientFilter explicitly
       //       if they set filter and want gradientFilter to be different.
-      config.gradientFilter =
-          (VKLFilter)this->template getParam<int>("gradientFilter", 
-              this->hasParam("filter") ? config.filter : config.gradientFilter);
+      config.gradientFilter = (VKLFilter)this->template getParam<int>(
+          "gradientFilter",
+          this->hasParam("filter") ? config.filter
+                                   : globalConfig.gradientFilter);
 
-      config.maxSamplingDepth =
-          this->template getParam<int>("maxSamplingDepth", config.maxSamplingDepth);
+      config.maxSamplingDepth = this->template getParam<int>(
+          "maxSamplingDepth", globalConfig.maxSamplingDepth);
     }
 
     template <int W>
@@ -43,8 +45,8 @@ namespace openvkl {
     {
       assert(attributeIndex == 0);
       CALL_ISPC(VdbSampler_computeSample_uniform,
-                this->grid,
-                &this->config,
+                volume->getGrid(),
+                &config,
                 &objectCoordinates,
                 static_cast<float *>(samples));
     }
@@ -58,8 +60,8 @@ namespace openvkl {
       assert(attributeIndex == 0);
       CALL_ISPC(VdbSampler_computeSample,
                 static_cast<const int *>(valid),
-                this->grid,
-                &this->config,
+                volume->getGrid(),
+                &config,
                 &objectCoordinates,
                 static_cast<float *>(samples));
     }
@@ -72,8 +74,8 @@ namespace openvkl {
     {
       assert(attributeIndex == 0);
       CALL_ISPC(VdbSampler_computeSample_stream,
-                this->grid,
-                &this->config,
+                volume->getGrid(),
+                &config,
                 N,
                 (const ispc::vec3f *)objectCoordinates,
                 samples);
@@ -88,8 +90,8 @@ namespace openvkl {
       assert(attributeIndex == 0);
       CALL_ISPC(VdbSampler_computeGradient,
                 static_cast<const int *>(valid),
-                this->grid,
-                &this->config,
+                volume->getGrid(),
+                &config,
                 &objectCoordinates,
                 &gradients);
     }
@@ -102,11 +104,24 @@ namespace openvkl {
     {
       assert(attributeIndex == 0);
       CALL_ISPC(VdbSampler_computeGradient_stream,
-                this->grid,
-                &this->config,
+                volume->getGrid(),
+                &config,
                 N,
                 (const ispc::vec3f *)objectCoordinates,
                 (ispc::vec3f *)gradients);
+    }
+
+    template <int W>
+    Observer<W> *VdbSampler<W>::newObserver(const char *type)
+    {
+      const std::string t(type);
+
+      if (t == "LeafNodeAccess") {
+        auto *obs = new VdbLeafAccessObserver<W>(*this, *volume->getGrid());
+        return obs;
+      }
+
+      return Sampler<W>::newObserver(type);
     }
 
     template struct VdbSampler<VKL_TARGET_WIDTH>;